+++ app/controllers/route53_aliaser/aliaser_controller.rb
@@ -1,0 +1,8 @@
+module Route53Aliaser
+  class AliaserController < ActionController::Base
+    def update
+      Route53Aliaser.update_alias_if_needed
+      head :ok
+    end
+  end
+end