+++ config/routes.rb
@@ -1,0 +1,3 @@
+Route53Aliaser::Engine.routes.draw do
+  root to: 'aliaser#update', via: [:get, :head]
+end