+++ lib/route53_aliaser/engine.rb
@@ -1,0 +1,5 @@
+module Route53Aliaser
+  class Engine < ::Rails::Engine
+    isolate_namespace Route53Aliaser
+  end
+end