+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# Route53Aliaser 0.0.2
+- add Rails Engine support
+- use Rails.cache when Rails is defined
+
 # Route53Aliaser 0.0.1
 - initial release